Refactor Themes & Tokens Structure
- make more accessible/visible by simlifying structure and bringing into `src` root.
