+++ gatsby-browser.js
@@ -11,7 +11,7 @@ import { SiteProvider } from './src/__context__/SiteContext';
 import { ShowProvider } from './src/__context__/ShowContext';
 
 // Import our themes
-import themes from './src/__ui__/themes';
+import themes from './src/__themes__';
 
 // Handle Application wrapper
 // eslint-disable-next-line
