+++ src/__themes__/primary/ui/buttons.js
@@ -1,0 +1,3 @@
+const buttons = {};
+
+export default buttons;
