+++ src/__themes__/primary/ui/index.js
@@ -1,0 +1,4 @@
+export { default as buttons } from './buttons';
+export { default as inputs } from './inputs';
+export { default as logo } from './logo';
+export { default as typography } from './typography';
