+++ src/__themes__/primary/ui/inputs.js
@@ -1,0 +1,3 @@
+const inputs = {};
+
+export default inputs;
