+++ src/__themes__/primary/ui/logo.js
@@ -1,0 +1,3 @@
+const logo = {};
+
+export default logo;
