+++ src/__themes__/primary/ui/typography.js
@@ -1,0 +1,3 @@
+const typography = {};
+
+export default typography;
