+++ src/__tokens__/index.js
@@ -1,7 +1,7 @@
 // logo elements
 
 // typography
-export { default as fonts } from './typography/typography';
+export { default as fonts } from './typography/fonts';
 export { default as typography } from './typography/typography';
 
 export { default as breakpoints } from './breakpoints';
