--- src/__ui__/themes/global.js
@@ -1,8 +1,0 @@
-// grayscale colors
-export const grayscale = {};
-
-// system colors
-export const system = {
-	black: '#000000',
-	white: '#FFFFFF',
-};
