--- src/__ui__/themes/primary/index.js
@@ -1,31 +1,0 @@
-import { grayscale, system } from '../global';
-import { primary, secondary, tertiary } from './_palette';
-import { ui } from './ui';
-
-export const theme = {
-	palette: {
-		primary: {
-			...primary,
-		},
-
-		secondary: {
-			...secondary,
-		},
-
-		tertiary: {
-			...tertiary,
-		},
-
-		grayscale: {
-			...grayscale,
-		},
-
-		system: {
-			...system,
-		},
-	},
-
-	ui: {
-		...ui,
-	},
-};
