--- src/__ui__/themes/primary/ui/index.js
@@ -1,11 +1,0 @@
-import './_buttons';
-import './_inputs';
-import './_logo';
-import './_typography';
-
-export const ui = {
-	buttons: {},
-	inputs: {},
-	logo: {},
-	typography: {},
-};
